Added Github Logos
This was made possible by HTML. If Markdown did not support HTML, This will not work !

Note:-
By default, the image would link to the raw file of the image. To prevent that, I linked it to https://github.com

Issues:-
The Github logos would not be visible if dark mode is enabled. In the website it will be visible because it has a white background. If you see the repository in dark mode, it will not be visible no matter what.